requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
apache-airflow==2.5.2
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.1.1
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
billiard==3.6.4.0
black==23.3.0
blinker==1.5
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.3
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.0.7
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==39.0.2
cycler==0.11.0
datasets==2.11.0
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
evaluate==0.4.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.11.0
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.3.3
flower==1.2.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.53.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.4
humanize==4.6.0
idna==3.4
inflection==0.5.1
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.9.0
libclang==16.0.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
ml-dtypes==0.0.4
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seqeval==1.2.2
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
tokenize-rt==5.0.0
tokenizers==0.13.3
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.5.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
xxhash==3.2.0
yarl==1.8.2
